feat(contents): add compare requirement builder
