+++ packages/engine/src/evaluators/development.ts
@@ -13,9 +13,7 @@ export const developmentEvaluator: EvaluatorHandler<
 	return ctx.activePlayer.lands.reduce(
 		(total, land) =>
 			total +
-			land.developments.filter(
-				(development) => development === id,
-			).length,
+			land.developments.filter((development) => development === id).length,
 		0,
 	);
 };
